rename vanguard assets
